chapter-10/streamlit/requirements.txt
streamlit
psycopg2-binary
sentence-transformers
pandas